+++ packages/ember-glimmer/tests/integration/components/will-destroy-element-hook-test.js
@@ -1,8 +1,6 @@
 import { set } from 'ember-metal';
 import { Component } from '../../utils/helpers';
 import { moduleFor, RenderingTest } from '../../utils/test-case';
-import { tryInvoke } from 'ember-utils';
-import { runAppend } from 'internal-test-helpers';
 
 moduleFor('Component willDestroyElement hook', class extends RenderingTest {
   ['@test it calls willDestroyElement when removed by if'](assert) {
@@ -32,26 +30,5 @@ moduleFor('Component willDestroyElement hook', class extends RenderingTest {
     assert.equal(willDestroyElementCount, 1, 'willDestroyElement was called once');
 
     this.assertText('');
-  }
-  ['@test still has access to this.$()'](assert) {
-    assert.expect(2);
-    let component;
-    let FooBarComponent = Component.extend({
-      tagName: 'div',
-      didInsertElement() {
-        component = this;
-      },
-      willDestroyElement() {
-        assert.ok(this.$(), 'willDestroyElement has access to element via this.$()');
-      },
-      didDestroyElement() {
-        assert.notOk(this.$(), 'didDestroyElement does not have access to element via this.$()');
-      }
-    });
-    this.registerComponent('foo-bar', { ComponentClass: FooBarComponent, template: 'hello' });
-    let { owner } = this;
-    let comp = owner.lookup('component:foo-bar');
-    runAppend(comp);
-    this.runTask(() => tryInvoke(component, 'destroy'));
   }
 });
